requirements.txt
PyQt6>=6.4.0
opencv-python>=4.8.0
imagehash>=4.3.1
moviepy>=1.0.3
librosa>=0.10.0
ffmpeg-python>=0.2.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0
tqdm>=4.65.0